backend/omni/pyproject.toml
[project]
name = "modai-backend"
version = "0.1.0"
description = "ModAI Chat Backend"
requires-python = ">=3.13"
dependencies = [
    "uvicorn",
    "fastapi",
    "pydantic",
    "pyyaml",
    "openai",
    "pyjwt",
    "python-dotenv",
    "sqlalchemy",
    "strands-agents",
    "strands-agents-tools",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "datamodel-code-generator[ruff]",
    "pytest",
    "pytest-asyncio",
    "ruff",
]

[tool.pytest.ini_options]
testpaths = ["src"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.uv]
package = true
